backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
strawberry-graphql[fastapi]==0.215.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
protobuf==4.25.1
websockets==12.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
